backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
requests>=2.31.0  # For API integrations
SSLCommerz-python==1.0  # For SSLCOMMERZ payment gateway
django-filter==23.3
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
drf-yasg==1.21.7 